+++ openpathsampling/pathsimulator.py
@@ -3,7 +3,7 @@ import logging
 from openpathsampling.util.todict import restores_as_stub_object
 import openpathsampling as paths
 from openpathsampling.pathmover import PathMover
-from ops_logging import initialization_logging
+from openpathsampling.util.ops_logging import initialization_logging
 
 logger = logging.getLogger(__name__)
 init_log = logging.getLogger('openpathsampling.initialization')
