+++ openpathsampling/util/ops_logging.py
